+++ karma.conf.js
@@ -22,6 +22,7 @@ module.exports = function(config) {
       'bower_components/restangular/dist/restangular.js',
       'bower_components/localforage/dist/localforage.js',
       'bower_components/angular-localForage/dist/angular-localForage.js',
+      'src/index.js',
       'src/**/*.js',
       'tests/**/*.spec.js',
       '.tmp/templates.js'